+++ Koronavilkku/Koronavilkku/ViewController/OnboardingViewController.swift
@@ -3,7 +3,7 @@ import UIKit
 import SnapKit
 import Combine
 
-class OnboardingViewController: UINavigationController {
+class OnboardingViewController: UINavigationController, UINavigationControllerDelegate {
     
     private var currentStep = 0
     private var button: RoundedButton? = nil
@@ -15,6 +15,7 @@ class OnboardingViewController: UINavigationController {
         // Note that the steps and StepId values must be in the same order.
         Step(id: .intro,
              buttonTitle: Translation.ButtonNext.localized,
+             showLanguageSelection: true,
              view: StepView(image: UIImage(named: "radar-static")!,
                             header: Translation.OnboardingIntroTitle.localized,
                             content: Translation.OnboardingIntroText.localized,
@@ -71,6 +72,7 @@ class OnboardingViewController: UINavigationController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        self.setDefaultStyle()
         self.setNavigationBarHidden(true, animated: false)
         let window = UIApplication.shared.windows.first
         let statusBarFrame = window?.windowScene?.statusBarManager?.statusBarFrame
@@ -82,8 +84,19 @@ class OnboardingViewController: UINavigationController {
             self.view.addSubview(blurEffectView)
         }
         
+        self.delegate = self
         self.view.backgroundColor = UIColor.Greyscale.white
         self.handleStartStep()
+    }
+    
+    func navigationController(_ navigationController: UINavigationController, willShow viewController: UIViewController, animated: Bool) {
+        if viewController is ChangeLanguageViewController {
+            self.largeTitleFont = .heading2
+            self.setNavigationBarHidden(false, animated: true)
+        } else if !navigationBar.isHidden {
+            self.largeTitleFont = .heading1
+            self.setNavigationBarHidden(true, animated: true)
+        }
     }
     
     private func performButtonAction(step: Step) {
@@ -244,6 +257,16 @@ class OnboardingViewController: UINavigationController {
             make.left.right.equalTo(viewController.view)
         }
         
+        if step.showLanguageSelection {
+            let languageButton = createLanguageSelectionButton()
+            scrollView?.addSubview(languageButton)
+            
+            languageButton.snp.makeConstraints { make in
+                make.top.equalToSuperview().offset(10)
+                make.right.equalToSuperview().inset(20)
+            }
+        }
+        
         let fade = FadeBlock()
         viewController.view.addSubview(fade)
         
@@ -319,6 +342,25 @@ class OnboardingViewController: UINavigationController {
                                  underline: false)
     }
     
+    private func createLanguageSelectionButton() -> UIButton {
+        let button = UIButton()
+        
+        button.addTarget(self, action: #selector(languageSelectionTapped), for: .touchUpInside)
+        button.setTitle("Language", for: .normal)
+        button.backgroundColor = UIColor.Secondary.blueBackdrop
+        button.setTitleColor(UIColor.Primary.blue, for: .normal)
+        button.titleLabel?.font = .bodySmall
+        button.layer.cornerRadius = 6
+        button.setImage(UIImage(named: "globe")?.withRenderingMode(.alwaysTemplate), for: .normal)
+        button.contentEdgeInsets = UIEdgeInsets(top: 6, left: 6, bottom: 6, right: 10)
+        button.imageEdgeInsets = UIEdgeInsets(top: 3, left: 0, bottom: 3, right: 0)
+        button.titleEdgeInsets = UIEdgeInsets(top: 0, left: 4, bottom: 0, right: -4)
+        button.imageView?.contentMode = .scaleAspectFit
+        button.tintColor = UIColor.Primary.blue
+
+        return button
+    }
+    
     private func hideScrollIndicator() {
         guard let indicator = scrollIndicatorButton else { return }
         self.scrollIndicatorButton = nil
@@ -346,23 +388,29 @@ class OnboardingViewController: UINavigationController {
             step(into: .intro)
         }
     }
+
+    @objc func languageSelectionTapped() {
+        pushViewController(ChangeLanguageViewController(), animated: true)
+    }
 }
 
 struct Step {
     let id: StepId
     let buttonTitle: String?
     let showScrollIndicator: Bool
-    let hasAcceptableExtraContent: Bool
+    let showLanguageSelection: Bool
     let view: StepView?
     let viewController: UIViewController?
     let requestPermission: (() -> ())?
     
-    init(id: StepId, buttonTitle: String? = nil, showScrollIndicator: Bool = false, hasAcceptableExtraContent: Bool = false, view: StepView? = nil, viewController: UIViewController? = nil, requestPermission: (() -> ())? = nil)
+    init(
+        id: StepId,
+        buttonTitle: String? = nil, showScrollIndicator: Bool = false, showLanguageSelection: Bool = false, view: StepView? = nil, viewController: UIViewController? = nil, requestPermission: (() -> ())? = nil)
     {
         self.id = id
         self.buttonTitle = buttonTitle
         self.showScrollIndicator = showScrollIndicator
-        self.hasAcceptableExtraContent = hasAcceptableExtraContent
+        self.showLanguageSelection = showLanguageSelection
         self.view = view
         self.viewController = viewController
         self.requestPermission = requestPermission
@@ -375,7 +423,6 @@ enum StepId: Int {
     case acceptTerms
     case enableApiInstructions
     case enableBluetooth
-    //    case allowNotifications
     
     var description: String { return "StepId(\(String(describing: self))" }
 }