+++ CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Localisation updates
 - Determine API disabled status from both exposureNotificationStatus and authorizationStatus
+- Corrected instructions on how to enable EN API on iOS 13.7+
 
 ### Added
 - English translation
 - Show licenses of used third party libraries inside the app, accessed from the Settings tab
+- Instructions on how to change the app language in Settings tab and Onboarding
 
 ## 1.0.2